+++ src/buffwriter.h
@@ -112,7 +112,7 @@ protected:
     std::string out_file_name(const std::string &string_pattern,
                               const std::string &fractal_type,
                               unsigned int bailout, unsigned int xrange,
-                              unsigned int yrange, unsigned int zoom,
+                              unsigned int yrange, double zoom,
                               unsigned int cores, double xcoord, double ycoord,
                               double z_real_min, double z_real_max,
                               double z_ima_min, double z_ima_max);
