+++ src/fractalzoom.h
@@ -40,7 +40,7 @@ public:
      * @param ycoord Y coordinate in the complex plane
      */
     void calcalute_zoom_cpane(double &xh, double &xl, double &yh, double &yl,
-                              unsigned int zoom, double xcoord, double ycoord,
+                              double zoom, double xcoord, double ycoord,
                               unsigned int width, unsigned int height);
 
 private:
